Rename data folder to model, Add controllers folder
